def simu.exprs
