+++ pyphs/numerics/tools.py
@@ -7,6 +7,7 @@ Created on Fri Jun  3 11:26:41 2016
 
 from sympy.printing.lambdarepr import lambdarepr
 import numpy
+import sympy
 import ast
 
 parser_sympy2numpy = {
@@ -45,6 +46,10 @@ def lambdify(args, expr, subs=None, simplify=True):
     """
 #    from sympy.printing.theanocode import theano_function
 #    return theano_function(args, expr)
+
+    if hasattr(expr, 'index'):
+        expr = sympy.Matrix(expr)
+
     if subs is not None:
         if hasattr(expr, 'index'):
             for i, e in enumerate(expr):
