+++ pyphs/simulations/config.py
@@ -11,9 +11,10 @@ standard = {'numtol': EPS,
             'load_options': {'decim': 1,
                              'imin': 0,
                              'imax': None},
+            'method': 'standard',
             'solver': 'standard',
             'fs': 48e3,
             'language': 'python',
             'timer': True,
-            'presubs': True,
+            'presubs': False,
             'split': True}
