+++ pyphs/simulations/simulation.py
@@ -4,9 +4,7 @@ Created on Tue May 24 11:20:26 2016
 
 @author: Falaize
 """
-from solvers.solvers import Solver
-from pyphs.misc.tools import norm
-from pyphs.numerics.tools import find
+
 from processes import process_py, process_cpp
 import config
 import time
@@ -37,47 +35,44 @@ class Simulation:
                                    'imax': None}
         """
 
-        # init configuration options
-        if opts is None:
-            opts = {}
+        # init config with standard configuration options
         self.config = config.standard
-        self.config.update(opts)
-
-        self.phs = phs
-
-    def init(self, sequ=None, seqp=None, x0=None, nt=None, opts=None):
+        # update with provided opts
         if opts is None:
             opts = {}
         self.config.update(opts)
-        self.init_phs()
-        self.phs.data.init_data(sequ, seqp, x0, nt)
-        init_args(self)
+        # store phs
+        self._phs = phs
 
-        # init args values with 0
-        setattr(self, 'args', [0, ]*self.phs.dims.args())
+###############################################################################
 
-        # init solver
-        self.solver = Solver(self)
-
-        # define state variation
-        self.phs.exprs.setexpr('dtx',
-                               [el*self.config['fs']
-                                   for el in self.phs.symbs.dx()])
-
-        # build numerical functions from functions in phs.exprs._names
-        self.phs.funcs.build()
-        init_funcs(self)
-
-    def init_phs(self):
+    def init_expressions(self):
         """
+        Init the phs.simu.exprs module that contains all expressions for \
+simulation.
         """
+
         if self.config['presubs']:
-            self.phs.apply_subs()
+            self._phs.apply_subs()
 
         # split system into linear and nonlinear parts
         force_nolin = not self.config['split']
+
         from pyphs.symbolics.structures.tools import split_linear
-        split_linear(self.phs, force_nolin=force_nolin)
+        split_linear(self._phs, force_nolin=force_nolin)
+
+        from expressions import SimulationExpressions
+        self.exprs = SimulationExpressions(self._phs)
+
+###############################################################################
+
+    def init(self, sequ=None, seqp=None, x0=None, nt=None, opts=None):
+        if opts is not None:
+            self.config.update(opts)
+
+        self.init_expressions()
+
+        self._phs.data.init_data(sequ, seqp, x0, nt)
 
     def process(self):
         """
@@ -92,6 +87,7 @@ class Simulation:
 
         if self.config['language'] == 'c++':
             process_cpp(self.phs)
+
         elif self.config['language'] == 'python':
             process_py(self)
 
@@ -104,111 +100,3 @@ class Simulation:
             time_ratio = time_it*self.config['fs']
             print 'ratio compared to real-time: {0!s}'.format(format(
                 time_ratio, 'f'))
-
-    def update(self, u, p):
-        """
-        update with input 'u' and parameter 'p' on the time step (samplerate \
-is numerics.fs).
-        """
-        # store u in numerics
-        self.set_u(u)
-        # store p in numerics
-        self.set_p(p)
-        # update state from previous iteration
-        self.set_x(map(lambda xi, dxi: xi + dxi, self.x(), self.dx()))
-        if self.phs.is_nl():
-            # update nl variables (dxnl and wnl)
-            self.update_nl()
-        # update l variables (dxnl and wnl)
-        self.update_l()
-
-    def update_nl(self):
-        # init it counter
-        it = 0
-        # init dx with 0
-        self.set_dxnl([0, ]*self.phs.dims.xnl())
-        # init step on iteration
-        step = float('Inf')
-        # init residual of implicite function
-        res = float('Inf')
-        # init args memory for computation of step on iteration
-        old_varsnl = [float('Inf'), ]*(self.phs.dims.xnl() +
-                                       self.phs.dims.wnl())
-        # loop while res > tol, step > tol and it < itmax
-        while res > self.config['numtol'] and step > self.config['numtol']\
-                and it < self.config['maxit']:
-            # updated args
-            self.iter_solver()
-            # eval residual
-            res = self.res_impfunc()
-            # eval norm step
-            step = norm([el1-el2 for (el1, el2) in zip(self.vnl(),
-                                                       old_varsnl)])
-            # increment it
-            it += 1
-            # save args for comparison
-            old_varsnl = self.vnl()
-
-    def update_l(self):
-        vl = self.update_lin().T.tolist()[0]
-        self.set_vl(vl)
-
-
-def init_args(simulation):
-    """
-    define accessors and mutators of numerical values associated to arguments
-    """
-    # generators of 'get' and 'set':
-    def get_generator(inds):
-        def get_func():
-            return [simulation.args[i] for i in inds]
-        return get_func
-
-    def set_generator(inds):
-        def set_func(lis):
-            for i in inds:
-                elt = lis.pop(0)
-                simulation.args[i] = elt
-        return set_func
-
-    # def lists of linear variables linvars
-    nxl = simulation.phs.dims.xl
-    nwl = simulation.phs.dims.wl
-
-    dic = {'vl': list(simulation.phs.symbs.dx()[:nxl]) +
-           list(simulation.phs.symbs.w[:nwl]),
-           'vnl': list(simulation.phs.symbs.dx()[nxl:]) +
-           list(simulation.phs.symbs.w[nwl:]),
-           'x': simulation.phs.symbs.x,
-           'dx': simulation.phs.symbs.dx(),
-           'dxnl': simulation.phs.symbs.dx()[nxl:],
-           'w': simulation.phs.symbs.w,
-           'u': simulation.phs.symbs.u,
-           'p': simulation.phs.symbs.p}
-
-    for name in dic:
-        _, inds = find(dic[name], simulation.phs.symbs.args())
-        setattr(simulation, name + '_symbs', dic[name])
-        setattr(simulation, name, get_generator(inds))
-        setattr(simulation, 'set_' + name, set_generator(inds))
-
-
-def init_funcs(simulation):
-    """
-    link and lambdify all funcions
-    """
-    # list of lambdified functions
-    simulation.funcs_names = simulation.phs.exprs._names
-
-    # generator of evaluation functions
-    def eval_generator(func, inds):
-        def eval_func():
-            args = (simulation.args[el] for el in inds)
-            return func(*args)
-        return eval_func
-
-    # link evaluation to internal values
-    for name in simulation.funcs_names:
-        func = getattr(simulation.phs.funcs, name)
-        inds = getattr(simulation.phs.funcs, name + '_inds')
-        setattr(simulation, name, eval_generator(func, inds))
