--- pyphs/simulations/solvers/__init__.py
@@ -1,1 +1,0 @@
-#
